move AnalysisStage class to workflow module
